CFilePath is refactored to inherit from QFileInfo
The CFilePath class in NDD was the primary means of abstracting file
path handling and manipulation in the authoring tool. The implementation
is heavily windows idiosyncratic. This change completely refactors the
class to rely upon Qt's file path handling abstractions: QFileInfo and
QDir. The goal is to continue to shrink this class until it can be
removed entirely and all call sites can use QFileInfo and friends
directly.

Task-number: QT3DS-221
Change-Id: I0e31c96982a3f3a9411346f6c1dd82a6d1fa00ca
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>