Remove CFilePath::GetCharStar()
This function returned pointer to an out-of-scope bytearray, causing
crashes. Removed the function entirely and fixed the places that used
it to properly scope the bytearray or just use toQString() instead.

Usage that was related to loading lua was simply removed as lua is
going away eventually anyway.

Task-number: QT3DS-438
Change-Id: Ifdcb28d445c156f04f361189f37132d56b6eede3
Reviewed-by: Adam Treat <adam.treat@qt.io>
Reviewed-by: Marianne Yrjänä <marianne.yrjana@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>