Fix hidden objects being visible on another slide
Hidden objects on one slide can appread visible on another, if they are
not used on that slide. This happens because the slide system invers the
active state at the end of the slide, which sets the hidden object as
visible. Since it is not used in the other slide, it won't be set
inactive.

Change the invert so that it only inverts the state if it was active.

Task-number: QT3DS-307
Change-Id: I159ef6d1a35a2de967da1432326deca0167174e7
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>