Widget viewer: Non-animating presentation does not scale correctly
There is a problem in the render state caching where a set hw state can be
different than a state in the cache, which results in hw states not being
set when they should. Add ResetStates function, which forces current
states into hw and call it in appropriate place.

Task-number: QT3DS-70
Change-Id: I626f4e4cd2e97f47776913a477a2f7c8647e2e29
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Marianne Yrjänä <marianne.yrjana@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>