Fix crash when undoing and redoing adding new slide at the end of list
Adding a slide as the last slide automatically sets the playthrough
property to 'previous' instead of its default value of 'next'.
This creates a HashMapSwapTransaction item for this property value.
When the redo of create slide happens, it appears to have lost this
property from the transaction's hash map of slide properties.
These properties are passed as a non-const reference, so I'm guessing
they get cleared during the undo to their default state.

Presumably missing properties from the hash map isn't a fatal error,
since after the assert the code simply inserts the missing property
into the map and things work fine.

Task-number: QT3DS-535
Change-Id: Ibc4607faee05e07ce055f63bec9da3e6d43b086d
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Marianne Yrjänä <marianne.yrjana@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>