Fix crash with fill/fit modes on large screens
QSize initializes to (-1, -1) by default, which resulted in negative
desired size. This was masked in most cases by the aligning size
to nearest four bytes, i.e. (0, 0), but if the fill/fit scaling
resulted in multiplier larger of 4, then the size would be aligned
to (-4, -4), leading to a crash.

Task-number: QT3DS-722
Change-Id: I042aec7237181840e49f9ed0312ee055407141c6
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>