Use temporary preview uip/uia if the presentation is dirty
Viewer now only saves a copy of the presentation to a temporary .uip
file when preview in a viewer is selected and the presentation is
dirty. A temporary .uia file is also made if one is found with the same
file stem. The temporary .uip and .uia files are removed if either
viewer or studio application exits.

Task-number: QT3DS-600
Change-Id: Ic4e0a3494f48f7831b9f8d956aa731c0e6a23b5d
Reviewed-by: Janne Kangas <janne.kangas@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>