Fix runtime shader cache
The shader cache has been broken (both at runtime and when persistence
is enabled) because a prefix for shader cache key was added when for
seemingly debug purposes. The result has been a missed lookup because
the key in the cache does not match the one we use for lookup.

Change-Id: Ib95cb0402979513c1516e5682054763d358ca42d
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>