Enable hdr image conversion to batch loader with es2
With es2 renderer the hdr images are converted to 32-bit rgb
since rgbf16 is not supported. Direct loader already does that, but
the batch loader needs to know the context version to convert
the hdr images to 32-bit rgba. Add the context version to the batch
loader to support the conversion.

Task-number: QT3DS-36
Change-Id: I8a7abda5ff1b79ced7d6531b46f6aa0661bfca16
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>