Fix massive memory leak when dragging object
Fix the memory leak that happens when object is dragged in view other than
scene view. This is temporary fix since memory is still leaked if the
active layer is changed while in view other than scene view. The core
problem is in the engine, because it doesn't release allocated resources
when the SEditCameraLayerTranslator is deallocated.

Task-number: QT3DS-974
Change-Id: I70dc69ef6167027037b5c8e8d22360db350b0351
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>