Fix scaling of fonts so that it is portable across screens
Rendering fonts on screens with different logical dpi was resulting in
inconsistent sizes of fonts in studio projects. This fixes an issue I
was seeing where my Linux host was displaying fonts at seemingly normal
size, but my QNX target with a different logical dpi had abnormally
large fonts. With this change, the rendering of the fonts looks
identical.

Change-Id: Ide4f0b507e5839808b5ea185fe346613a54264a2
Reviewed-by: Antti Määttä <antti.maatta@qt.io>
Reviewed-by: Janne Koskinen <janne.p.koskinen@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Pasi Keränen <pasi.keranen@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>