Optimize iterating through child objects
GetAssetChildrenInTimeParent was getting called unnecessarily much,
as it was called every time GetChildrenCount or GetChild was called.
According to profiling it was responsible for vast majority of the time
spent when changing a slide. Optimized various loops that iterate
through object's children to only call this function once.

Task-number: QT3DS-1601
Change-Id: Ia006acb76bd818c390a36f6fe9f1656660974c0e
Reviewed-by: Antti Määttä <antti.maatta@qt.io>
Reviewed-by: Jere Tuliniemi <jere.tuliniemi@qt.io>
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>