Clear prev/next sibling when adding a child node
Normally prev/next sibling are cleared when the node is removed from
the previous parent, but apparently having layer as parent is indicated
by a null parent, so the remove is not called when moving layer's
direct children. They will retain old prev/next sibling
information, which can corrupt other child nodes of the same parent
when the original node is moved out in a future transaction.

Task-number: QT3DS-1551
Change-Id: I6203789f1e7cf45d60a8358f3c71c2deff07fdd7
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>