Fix image probe FOV calculation
Fix FOV equations. Also, fix case for FOV of 180 degrees where image
lighting map was used in wrong orientation. With this fix, the negative z
axis passes through the center of the lighting map, with edges of the image
wrapped around the scene, i.e. with default camera settings the center of
image is in the direction of camera. This holds true for both cases of FOV = 180 and
FOV < 180 degrees.

This fix must be replicated on runtime side for results to match.

Task-ID: QT3DS-154

Change-Id: I917f4f1897e59294aa369933fc9b10ae849952e6
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>