Fix adjusting timebar by dragging start/end handles
In case of overlapping start and end handles, now choose the handle
depending on the click position relative to start of the bar.
Never select start handle if end time is zero.
Also allow adjusting the duration to actual zero instead of just very
close to it.

Task-number: QT3DS-1869
Change-Id: I57d06c6e70bb22fd320bcd01adf021a15f13caf6
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>