Fix startup dialog flickering when dragging to another screen
It's still unclear what the actual issue was, but copying paintEvent
handling from about dialog, and making the dialog form more like
about dialog, i.e. adding a layout to the dialog and
setting one of the QLabels to have word wrap fixes the flickering.
The fact that word wrap makes a difference likely means
that the root cause is some Qt bug and this might break again when
Qt version changes.

Task-number: QT3DS-939
Change-Id: If75f5e1d9ed57e3f72b13209a9d5fe6878371b34
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>