Fix undo/redo of sliders
Separated the value from the backend and the value adjusted by the
user so that we will never break the binding. Also renamed the
signals to more descriptive of what is actually happening and
fixed the action panel property sliders to only change the value
when it is committed.

Task-number: QT3DS-1895
Change-Id: I4377f741ddbeef8a8522124a22ef9dd8106aafca
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>