Make keyframe UI update asynchronous
Updating keyframes to UI is done by first deleting all keyframes of the
affected property and then recreating them from binding. If this is
done synchronously, binding is not always up to date when the change
affects more than one keyframe, as notifications are done per-keyframe
basis. This causes crashes e.g. when multiple keyframes are deleted.

Changed the UI to be updated asynchronously to ensure binding is
always up to date. In addition to fixing the crash, this also makes
make multi-keyframe operations more efficient, as the keyframes are
only recreated once per affected property row.

Task-number: QT3DS-1963
Change-Id: I18ceca36b9bc8c24a49787727f76d0762cb73df3
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Janne Kangas <janne.kangas@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>