Add presentation id to the presentation settings panel
- The presentation settings panel appear after creating a new
 presentation
- The presentation settings panel now show the presentation Id (for the
currently open presentation)
- Added logic to make sure the presentation id remains unique
- Added a command prompt option (--add) to add a presentation to project,
if the option is absent, a new project is created
- Added a context menu 'edit presentation id' action to the presentation
items in the project palette
- removed the sub presentations menu option and dialogue
- Added the ability to import a presentation. This will also import all
presentations related assets (images, models, fonts...etc). If an asset
already exists a message box is displayed so that the user can choose
whether to override or skip the file. The only missing thing is that
subpresentations are not imported. This part is to be discussed and/or
implemented in a separate task.

Change-Id: I9fa96227efa82b4cef90615f1f069530419db7b6
Reviewed-by: Antti Määttä <antti.maatta@qt.io>