Fix preview view issues
Dragging was inaccurate when preview view was shown due to incorrect
translation being left after preview view was rendered. Changed the
preview view to render first to a texture, which is then draw on top
at the end of the regular view rendering pass.

The viewport for camera was also incorrect for preview view.

Task-number: QT3DS-2036
Change-Id: I3ed8f092c2a2a2f3806d5219ca8fb7718a8d4fe0
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>