Import Qml streams
Allow importing visual Qml files (Qml streams) to the Studio. visual Qmls
are detected and put in the 'qml streams' folder. the qml file is
parsed recursively and any node with a valid 'source' attribute is
imported as well. Imported Qmls are added to the project file for use as
a sub-presentation.

Task-number: QT3DS-2061
Change-Id: Ic0ac0178adb4bac06972a9e553c1dc0bb6c5286d
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Janne Kangas <janne.kangas@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>