Do not query sibling timerange when creating new object by DnD
During object creation, to-be-siblings are queried for start/end times so
that the new object start/end can be set to same values. For certain types
of siblings (Component), these properties do not exist at this point of
time. Fix moves time range setting to a later time when it can be queried
from parent.

Change-Id: If93465ba24784654533b0d0f126cc134ee41b988
Task-id: QT3DS-2320
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>