Remove renderable clearing on .uia file update
Manual editing of .uia files is no longer supported, so removed the
code to detect that. Updating the in-memory renderables in use should
be done explicitly when subpresentation ids change (or when
subpresentations are deleted, though that is not currently supported.)

Also fixed a bug in updating subpresentation id to .uip files.

Task-number: QT3DS-2134
Change-Id: I3a239523d88c66c60c8cfd0b52d949c60fa7033f
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>