Move the material container to be under the Scene
Now that the latest runtime2 supports materials outside the layers,
have the material container be under the Scene object. In addition, make
the container object itself be a material object as preferred for
runtime2. Object reference lookup also needs an exception for materials
with child objects for the editor.

Task-number: QT3DS-2108
Change-Id: I7c54574a9a9a3a98c026ae61659a99aef8b70ebd
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>