Fix duplication and renaming of materials
Save the original filepath to the material file to check if the original
file still exists. If it exists it is a duplicate. If not, treat it as a
rename action and rename the hidden material object. Material file
watching is also removed from DocumentEditor altogether and everything
is handled in InspectorControlModel::setMatDatas.

Task-number: QT3DS-2196
Task-number: QT3DS-2194
Change-Id: Ie3afdadef83baacdc8a5eb887653a67e64511732
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>