Fix welcome screen glitches
The dummy label is needed to fix background quadrupling issue when
dragging the welcome screen to another screen with different
device pixel ratio. It is unclear why exactly it fixes it,
and why it needs to be a label with some visible text to do so,
instead of some other widget. It is somehow related to the layout
handling on the dialog.

Task-number: QT3DS-2392
Change-Id: I945d29e6d735d400c80668cab458bf46e9b57f0c
Reviewed-by: Kimmo Leppälä <kimmo.leppala@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>