Fix presentation reload
When loading presentation after another was already loaded, the
presentation was not shown correctly or it disappeared after a while.
First problem was caused by translator not being created for the
presentation and the second one because runtime2 slide player was running
and hiding the objects after slide ended.

Task-number: QT3DS-2071
Change-Id: I8390568823fe7e17a8d231a26729eac09ca6020a
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>