Refactor mouse handling in project palette
Due to regression QTBUG-70999, mouse responsiveness on project palette
was severely hampered. Refactored the mouse handling code to get rid
of a need for the hacky doubleclick timer to improve things a bit,
but it's still missing the first click after a double click. We can't
fix that in our code.

Also fixed an issue about opening the edit dialog for presentation id
and prevented trying to drag presentations that do not have id.

Change-Id: I2269a7adfeb9917ed5fa2be11e0356ef58b0d8f6
Fixes: QT3DS-2450
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Janne Kangas <janne.kangas@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>