Add generic import asset type for property datatype queries
Adds "controllable" type that can be used to query all property datatypes
without need to specify the object type that the property is attached to.
Existing SImportAsset class does not provide iteration macros for all
needed properties, resulting in crashes when datainput control bindings
in subpresentations are checked for datatypes.

Change-Id: I670432f2b9089c65a72611328c508e130a7970be
Task-id: QT3DS-2515
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>