Fix crash on exit
Saving preferences failed on macOS, causing the crash.
Now preferences are saved each time preference dialog
is closed, or a recent item is added or changes.
Also changed the preference folder into a more
describing one instead of the old Qt3DSComposer, which
is not used anywhere else in code or docs.

Task-number: QT3DS-2499
Change-Id: I1f725e246c3fd491aa9600ddf0503e96af6ed1de
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>