Fix datainput chooser popup alignment
Introduced new options for CDialogs::showWidgetBrowser alignment:
Combobox, ToolButton, or Center. Combobox is the default and positions
the popup as it has before, which is suitable for comboboxes.
ToolButton alignment positions the corner of the popup over the
toolbutton and prefers to open the popup above rather than below.
Center alignment simply centers the popup on given point.

Task-number: QT3DS-2504
Change-Id: Ibca6895010baec19508259eb6b3976b6082740e9
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Jere Tuliniemi <jere.tuliniemi@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>