Ensure browser popups always end up with desired size
Showing a browser popup on a different screen than where it was shown
previously can cause it to show up incorrectly sized. Since the
involuntary resizing is done somehere inside Qt's handling of show,
simplest fix is to just ensure that once the popup is shown,
it is resized to correct size. This means the popup will briefly
show up incorrectly sized, but since this only happens rarely
(once after editor is moved to a different screen), this seems
like acceptable solution.

Task-number: QT3DS-2516
Change-Id: I80af3071dd099758730d971c987dadafca3583bd
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Jere Tuliniemi <jere.tuliniemi@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>