Fix issues with rendering layers into overlay preview
The renderer gets confused when we try to fake the presentation
size, so always render the preview as presentation size instead of
trying to render it smaller. We already need full size preview for
the scene camera tab, so usually not rendering an additional smaller
version is a performance gain. Since the overlay size is now capped to
the presentation size, this shouldn't affect the preview quality.

Task-number: QT3DS-2568
Task-number: QT3DS-2575
Task-number: QT3DS-2641
Change-Id: I75d4ab08089998c2b718ff9f9ab2640262997789
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Janne Kangas <janne.kangas@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>