Integrate Google breakpad with Qt 3D Studio
When a crash happens, a minidump file is generated and a friendly popup
appears that collects info and send it along the crash minidump file to
the server. Sentry.io is used as the crash server.
Breakpad is referenced via an environmental variable (BREAKPAD_SOURCE_DIR)

Task-number: QT3DS-2522
Task-number: QT3DS-2523
Change-Id: I1fd53b9bfea4512ef74abc2a43834ad17224f561
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>