Fix rename and update refmats upon rename
Implemented rename in a way that doesn't need presentation or filename
information from the matdef file. Absolute path saved in the matdef file
is compared to the absolute path of the material in the current
presentation. If the paths are the same and don't exist anymore in the
filesystem, a rename has occurred. The internal material and any
referenced materials in the scene are renamed.

Task-number: QT3DS-2108
Change-Id: I77951cff902e243ef3e8596d6946c74a0d9ad33a
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>