Fix bugs concerning freshly created matdefs
Material creation inside the material container has to be a transaction
so that the created material functions correctly. This creates an
undoable command which for now is just removed from the undo stack after
the material creation.

Some undo behaviors for the matdefs are also fixed.

Task-number: QT3DS-2685
Task-number: QT3DS-2686
Change-Id: I5abe9884745b7092ff454e2463b74898ab87822a
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>