Force material container to have zero duration
Since the material container is a material itself, it doesn't get
saved with endtime, so we force set it to zero on all slides so it
doesn't affect the overall time of the presentation animations.

This fixes the issue of dragging the playhead past the end of timeline
increasing the time up to default 10 seconds.

Task-number: QT3DS-2606
Change-Id: Ie4d91da63c307baf7cbe3cc0c02991d8aca56ad0
Reviewed-by: Jere Tuliniemi <jere.tuliniemi@qt.io>
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>