Fix missing glsllib errors
The old system assumed the glsllibs are in the filesystem.
Instead, we have them in a resource file. We do not have separate
versions for different OpenGL versions, so we can get rid of the
OpenGL version in the search path, hence allowing the glsllibs to be
found in the resource file.

Task-number: QT3DS-2798
Change-Id: I03cbb123cf4555456fc11423a82498703b07131d
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>