Use original presentation id when importing a presentation
If an original presentation id can be resolved for an imported
presentation, that will be used as the id instead of generating a new
one. In case the id is a duplicate with an existing id, then a new one
will be generated based on the file name.

Task-number: QT3DS-2486
Change-Id: I1bf75fedfa9230a4f73271522800eaab02942ab4
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Janne Kangas <janne.kangas@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>