Add metadata field to datainput definition
Add a single key-value metadata to datainput item definition. Add editable
field in datainput dialog for entering metadata.

Metadata is provided as convenience for the user integrating
Q3DSPresentation with external code. It can be used to better identify the
intended use for a specific datainput, for example. Metadata is stored in
UIA (project) file. Corresponding getter function will be implemented
on Runtime side.

Change-Id: I60d4f8318a7552bd82b41c857d4314be0828f013
Task-id: QT3DS-2992
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>