Fix StudioUtils::devicePixelRatio
StudioUtils::devicePixelRatio was not using the correct window
for checking the pixel ratio. Now the window can be specified,
and if it is not, then main frame will be used.

Also, there were two separate FixedDevicePixelRatio methods for
getting correct pixel ratio. Removed those and fixed the
implementation of the StudioUtils::devicePixelRatio accordingly.

Task-number: QT3DS-2901
Change-Id: Icd95dbd006d6fbce40056961a33a01cb0651ef53
Reviewed-by: Antti Määttä <antti.maatta@qt.io>