Port visual aid widget
Draws visual aid wireframes and billboard icons. Models used for the
wireframes are temporary ones.

The foreground layer is split into two layers. One for just showing
the widget models and one for models that are meant to be picked.
Layer settings are also now copied to the foreground layer so that
the widgets retain the same proportions as the objects in the
selected layer.

Another issue that remains is that the object matrices are one frame
behind so icons etc. lag behind movement.

Task-number: QT3DS-2910
Change-Id: I0c4b85c94e097e26a5364e8acbd9fa39bd74cbe5
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>