Port selection bounding boxes
Red bounding boxes are now drawn for selected objects. The model used
for the bounding box is a temporary one causing unnecessary lines to be
drawn. Bounding box calculation also can take time for models with large
amount of vertices so that should be done before model selection in the
future.

Q3DSSelectionWidget name is used for the new widget handling the
bounding boxes. The widget handling translation, scale and rotation is
renamed Q3DSManipulationWidget.

Task-number: QT3DS-3043
Change-Id: I243a22d5dc883c8719a66347173d54c75da267d5
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>