Fix cursor issues with drag adjusting float fields
Now properly account for the screen when setting the mouse position.
Also use the center of the parent widget for reference point instead
of the center of the main screen. This fixes an issue where cursor
sometimes resets incorrectly to what the scene view needs instead of
the default arrow after the drag.

Task-number: QT3DS-3144
Change-Id: Ia1b5c5244ae89bcc57ccd946fbb007beb40a8975
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>