Add dynamic object creation to presentation C++ API
A new model element can be dynamically created into the scene.
The slide and properties of the element can be specified at
creation time. A material element is automatically created for
each added model. The material is specified with a custom
"material" property in the list of creation properties.

Task-number: QT3DS-3209
Change-Id: I52b0e929023092110820405473a9399b67a305cd
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>