Fix crash in editor when using qml stream
The crash occurs due to stack overflow due to recursive subpresentation
initialization in runtime codes. This happens because the qml
subpresentation has the same id as the presentation filename, which is
assigned to default initial id from the presentation filename. When the
layer, which is using the qml subpresentation, is initialized in the
runtime, it gets the uip presentation instead of the qml subpresentation.
The same begins to happen recursively.

Add the ability to change the initial presentation id from the surface
viewer api, which is used to render the subpresentations in the editor so
that it gets valid id instead of the id from filename.

Task-number: QT3DS-3345
Change-Id: I45f3a8ce9b5dc030f120baa1a8aed6c284ea7e3c
Reviewed-by: Jere Tuliniemi <jere.tuliniemi@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>