Allow subpresentations as textures for custom materials and effects
Effects already supported taking the subpresentation as a texture in
the runtime. That behavior is now copied by the custom materials.

This also fixes a runtime bug where the subpresentation was rendered
only if a default material already had the subpresentation as a texture.

Now custom material and effect properties are iterated prior rendering
If they contain subpresentations, those are added the to the render task
list.

Task-number: QT3DS-3165
Change-Id: I7bddec2fece442065d517320d0e381eba5f1b159
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>