Clarify material creation/deletion API
Replaced element path with presentation ID in createMaterial, as that
is what we are actually interested in. Removed element path from
deleteMaterial, the presentation ID is expected in the name.
The materialsCreated signal will now include the presentation ID in
the material name if the material is in a subpresentation.

Task-number: QT3DS-3623
Change-Id: I0c38c0eb52f85005256c3bf3dee45be384d66d37
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Jari Karppinen <jari.karppinen@qt.io>
Reviewed-by: Janne Kangas <janne.kangas@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>