Stop adjusting properties if drag plane intersection goes behind camera
When doing widget drags, depending on the widget orientation, the drag
plane can be such that mouse cursor position and drag plane
intersection can sometimes be behind the camera. These cases manifest
in sudden flipping of sign of dragged property values, which is not
desirable. The issue is fixed by preventing property adjustment when
intersection is behind camera.

Task-number: QT3DS-3521
Change-Id: I25a3b73e9e0dc819be22b0585013c49c7046e2b1
Reviewed-by: Janne Koskinen <janne.p.koskinen@qt.io>