Fix material duplication behavior
Disable copying of referenced material properties and if somehow the
user is able to duplicate a referenced material that points to a
referenced material, do a recursive lookup for the original material.

If an animatable material is duplicated, it's type is changed to
referenced material which points to the duplicated material in the
container. Or in other words it is changed to a basic material.

Task-number: QT3DS-3679
Task-number: QT3DS-3973
Change-Id: Ic09bb897d24f7c10984c5cc2f80556be844d9c51
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>