Ensure that only one camera is active per layer
Only allow one camera to be active on a layer. Activating a camera hides
the previously active one. In case of multi-selection being activated,
pick one of layer cameras in the selection and make only it active.

Also, when loading a presentation, inactivate all but one camera per layer
and show user a popup.

Change-Id: I75768f737d22f52bc94f16f10869486f7c2cad7c
Task-id: QT3DS-2153
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>